Remove Yarn
The TigerBeetle Node client has no dependencies, apart from developer
devepencies, so Yarn provides less value. We also hit a Yarn-specific
bug [1] that was not backported, and later versions of Yarn have
become more complex with the additional indirection of Corepack, which
is still an experimental Node.js API.

Also deduplicate the Zig library build step, so that there is only
one canonical build command, not another also in postinstall.sh. This
caused us some debugging trouble in the past, when a different command
was being used to build the library, instead of that in package.json.

[1] https://github.com/yarnpkg/berry/issues/2232#issuecomment-818514929